Updated components using SCSS modules to rename files .module.scss for new webpack config
