+++ client/src/components/Navigation/NavigationItems/NavigationItem/NavigationItem.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {NavLink} from 'react-router-dom';
 
-import classes from './NavigationItem.scss';
+import classes from './NavigationItem.module.scss';
 
 const navigationItem = (props) => (
 	<li className={classes.NavigationItem}>
