+++ client/src/hoc/Layout/Layout.js
@@ -6,7 +6,7 @@ import Aux from '../Aux/Aux';
 import Toolbar from '../../components/Navigation/Toolbar/Toolbar';
 import SideDrawer from '../../components/Navigation/SideDrawer/SideDrawer';
 
-import classes from './Layout.scss';
+import classes from './Layout.module.scss';
 
 @inject('stores')
 @observer @autobind
